comitted revised dust_board package, changed the wget to selarium. Added Meteogram for 5 Locations

+++ dust_board/__wget.py
@@ -4,6 +4,13 @@ from urllib.parse import urljoin
 
 import requests
 from bs4 import BeautifulSoup
+from selenium import webdriver
+from selenium.webdriver.firefox.service import Service
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from webdriver_manager.firefox import GeckoDriverManager
+from selenium.webdriver.firefox.options import Options
 
 
 def HttpsScan(url):
@@ -16,6 +23,35 @@ def HttpsScan(url):
     :rtype: List[str]
     """
 
+    # Setup Firefox options for headless mode
+    options = Options()
+    #options.headless = True
+    options.add_argument('--headless')
+
+    # Initialize the Firefox driver
+    driver = webdriver.Firefox(service=Service(GeckoDriverManager().install()), options=options)
+    driver.implicitly_wait(10)
+
+    # Open the URL
+    driver.get(url)
+
+    # Wait for the page to load completely
+    try:
+        WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.ID, "teleports")))
+
+        # Find all links in the directory listing
+        links = driver.find_elements(By.TAG_NAME, "a")
+
+        # Extract the file names from the links
+        urls = [link.get_attribute('href') for link in links if link.get_attribute('href') and not link.get_attribute('href').endswith('/')]
+        files = [os.path.basename(url) for url in urls]
+        return urls, files
+
+    finally:
+        # Close the browser
+        driver.quit()
+
+    '''
     try:
         response = requests.get(url)
         if response.status_code == 200:
@@ -30,7 +66,7 @@ def HttpsScan(url):
     except Exception as e:
         print(f"Error: {e}")
         return []
-
+    '''
 
 def RequestWebFile(file_url, fname, directory):
     r"""
@@ -50,37 +86,37 @@ def RequestWebFile(file_url, fname, directory):
     assert type(file_url) == type(fname) == type(directory) == str, "Input should be string"
 
     # Define the header setup for downloading the files
-    headers = {
-        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36',
-        'Referer': 'https://example.com'
-    }
+    headers = {'X-Client-Token': 'Qg9Z4I0WWinLKQVYjVQw2MKLYrmXVXh5-0DiVEzl6rA'}
 
     target_path = os.path.join(directory, fname)
 
-    response = requests.get(file_url, headers)
-    if response.status_code == 200:
-        with open(target_path, 'wb') as f:
-            f.write(response.content)
+    if os.path.exists(target_path) == False:
+        response = requests.get(file_url, headers=headers)
+        if response.status_code == 200:
+            with open(target_path, 'wb') as f:
+                f.write(response.content)
 
-        # Check if the downloaded file is a bz2 file
-        if target_path.endswith('.bz2'):
-            try:
-                # Open the downloaded file and decompress it
-                with open(target_path, 'rb') as f_bz2:
-                    decompressed_content = bz2.decompress(f_bz2.read())
-                # Write the decompressed content back to the file
-                with open(target_path[:-4], 'wb') as f_decompressed:
-                    f_decompressed.write(decompressed_content)
-                # Remove the bz2 file
-                os.remove(target_path)
+            # Check if the downloaded file is a grib2 file
+            if target_path.endswith('.grib2'):
+                try:
+                    ## Open the downloaded file and decompress it
+                    #with open(target_path, 'rb') as f_bz2:
+                    #    decompressed_content = bz2.decompress(f_bz2.read())
+                    ## Write the decompressed content back to the file
+                    #with open(target_path[:-4], 'wb') as f_decompressed:
+                    #    f_decompressed.write(decompressed_content)
+                    ## Remove the bz2 file
+                    #os.remove(target_path)
+                    return True
+                except Exception as e:
+                    print(f"Error decompressing file: {e}")
+                    return False
+            else:
                 return True
-            except Exception as e:
-                print(f"Error decompressing file: {e}")
-                return False
         else:
-            return True
+            return False
     else:
-        return False
+        return True
 
 
 def ModelLoad(model, parameters_dict):
@@ -104,9 +140,10 @@ def ModelLoad(model, parameters_dict):
         url = urljoin(env_dict.model_url[model],
                       env_dict.__getattribute__((model + '_variable'))[vname])
         fileurls, filenames = HttpsScan(url)
+        print(fileurls)
         # Remove ../ path from scaned list
-        fileurls = fileurls[1::]
-        filenames = filenames[1::]
+        fileurls = fileurls[3:]
+        filenames = filenames[3:]
         for i in range(len(fileurls)):
             status = RequestWebFile(fileurls[i], filenames[i], local_storage_path)
             assert status == True, "Web request failed"
